RGBD-Face-Avatar-GAN/requirements.txt
numpy==1.18.3
matplotlib==3.2.1
opencv-python==4.2.0.34
python-osc==1.7.4
Pillow == 7.2
open3D~=0.11.2
tqdm~=4.54.1
face-alignment
torchsummary~=1.5.1
tensorboard
onnx~=1.8.0
dlib
cmake